Created implementation of CreateBooking and SaveBooking. Tests dont work yet
