+++ pom.xml
@@ -59,7 +59,7 @@
         <dependency>
             <groupId>dk.teamc</groupId>
             <artifactId>LSD-Case3</artifactId>
-            <version>1.0.3</version>
+            <version>1.0.6</version>
         </dependency>
         <dependency>
             <groupId>com.h2database</groupId>
