--- src/main/java/backend/Main.java
@@ -1,18 +1,0 @@
-package backend;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.ConfigurableApplicationContext;
-import org.springframework.context.annotation.AnnotationConfigApplicationContext;
-
-@SpringBootApplication
-public class Main {
-    public static void main(String[] args) {
-        ConfigurableApplicationContext context = SpringApplication.run(Main.class, args);
-
-
-
-        Test test = context.getBean(Test.class);
-        test.insertIntoDB();
-    }
-}
