--- src/main/java/backend/datalayer/constants/Place.java
@@ -1,7 +1,0 @@
-package backend.datalayer.constants;
-
-public enum Place {
-    NONE,
-    HOTEL,
-    AIRPORT
-}
