--- src/main/java/backend/datalayer/dao/AddressRepository.java
@@ -1,9 +1,0 @@
-package backend.datalayer.dao;
-
-import backend.datalayer.entity.AddressDB;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AddressRepository extends JpaRepository<AddressDB, Long> {
-}
