--- src/main/java/backend/datalayer/dao/AirportRepository.java
@@ -1,9 +1,0 @@
-package backend.datalayer.dao;
-
-import backend.datalayer.entity.AirportDB;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AirportRepository extends JpaRepository<AirportDB, String> {
-}
