--- src/main/java/backend/datalayer/dao/BookingRepository.java
@@ -1,9 +1,0 @@
-package backend.datalayer.dao;
-
-import backend.datalayer.entity.BookingDB;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface BookingRepository extends JpaRepository<BookingDB, Long> {
-}
