--- src/main/java/backend/datalayer/dao/CustomCarRepository.java
@@ -1,9 +1,0 @@
-package backend.datalayer.dao;
-
-import booking.dto.BookingCriteria;
-import booking.dto.CarSummary;
-import java.util.Collection;
-
-public interface CustomCarRepository {
-    Collection<CarSummary> findAvailableCars(BookingCriteria bookingCriteria);
-}
