--- src/main/java/backend/datalayer/dao/HotelRepository.java
@@ -1,9 +1,0 @@
-package backend.datalayer.dao;
-
-import backend.datalayer.entity.HotelDB;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface HotelRepository extends JpaRepository<HotelDB, Long> {
-}
