+++ src/main/java/booking/Main.java
@@ -1,0 +1,16 @@
+package booking;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.ConfigurableApplicationContext;
+
+@SpringBootApplication
+public class Main {
+    public static void main(String[] args) {
+        ConfigurableApplicationContext context = SpringApplication.run(Main.class, args);
+
+        Contract contract = context.getBean(Contract.class);
+        /*Test test = context.getBean(Test.class);
+        test.insertIntoDB();*/
+    }
+}
