+++ src/main/java/booking/datalayer/constants/Place.java
@@ -1,0 +1,7 @@
+package booking.datalayer.constants;
+
+public enum Place {
+    NONE,
+    HOTEL,
+    AIRPORT
+}
