+++ src/main/java/booking/datalayer/dao/AirportRepository.java
@@ -1,0 +1,9 @@
+package booking.datalayer.dao;
+
+import booking.datalayer.entity.AirportDB;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AirportRepository extends JpaRepository<AirportDB, String> {
+}
