+++ src/main/java/booking/datalayer/dao/BookingRepository.java
@@ -1,0 +1,9 @@
+package booking.datalayer.dao;
+
+import booking.datalayer.entity.BookingDB;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface BookingRepository extends JpaRepository<BookingDB, Long> {
+}
