+++ src/main/java/booking/datalayer/dao/HotelRepository.java
@@ -1,0 +1,9 @@
+package booking.datalayer.dao;
+
+import booking.datalayer.entity.HotelDB;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface HotelRepository extends JpaRepository<HotelDB, Long> {
+}
