+++ src/main/java/booking/servicelayer/util/HelperFunctions.java
@@ -1,0 +1,9 @@
+package booking.servicelayer.util;
+
+public class HelperFunctions {
+    public static void nullCheck(Object o) {
+        if(o == null){
+            throw new NullPointerException(o.getClass().getSimpleName() + " must not be null");
+        }
+    }
+}
