+++ src/main/resources/application.properties
@@ -1,5 +1,3 @@
-
-
 spring.datasource.url=jdbc:mysql://localhost:3306/mysql
 spring.datasource.username=root
 spring.datasource.password=admin
